requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
pydantic==1.9.0
sqlmodel==0.0.6
psycopg2==2.9.3
alembic==1.7.7
passlib==1.7.4
python-jose==3.3.0
pytest==7.1.2
requests==2.27.1